+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Added `explain_policy_shards` to eventually replace `explain` attribute on `PolicyShard` class entirely.
 - Added `__contains__` to `EffectiveARP` classes
 - Added `reverse` method to `Condition` to reverse the operator/value to reverse the effect of the condition.
+- Added `factory` method to `PolicyShard` that uses `Condition.reverse` to try to eliminate `not_conditions` where possible.
+
 # 0.5.0
 
 - Renamed `dedupe_policy_shards` to `dedupe_policy_shard_subsets` to differentiate it from `delineate_intersecting_shards`.